Change the default hashing algorithm to SHA512
Note that SHA1 is not considered insecure when used for signing, see
https://github.com/pallets/itsdangerous/issues/147#issuecomment-609045879. However
seeing SHA1 often causes concern and upgrading to SHA512 is not that
painful. The cookie value is longer, but likely way below the cookie
size limit (unless users use a very long auth id).

This will break any existing cookies though - which I hope is ok as
this is a relatively new library.

This also follows Flask-Login that has used SHA512 for the past 6
years
https://github.com/maxcountryman/flask-login/commit/83eee57c817967e26bbdc38bb1971cdcbc158516.