Add a bearer mode
This allows Quart-Auth to be used to authenticate APIs designed for
non-webbrowser usage i.e. based on a token present in a bearer
authorization header.

This requires the tokens to be timed, as the API usage has no cookie
lifespan to set. This means that any existing tokens will be
considered invalid and hence this upgrade will require all existing
users to log back in.