Support multiple auth schemes
This is very useful when having say staff and admin logins, or
supporting different domains. The globals are still supported, and
will point to the QuartAuth instance marked as the singleton, as this
is a very common and useful case.

I've decided to warn, rather than error if multiple QuartAuth
instances share names as there are legitimate use cases for
this. However, it does pose an obvious security risk hence the
warning.

The TestClient mixin was probably always a mistaken design as the new
global functions are clearer and don't require silencing mypy errors.